adapted main guard
