+++ chapter 4/visual_elements/pages/job_selector.py
@@ -1,8 +1,5 @@
 import taipy.gui.builder as tgb
-from orchestration import auto_scenario
 from taipy.gui import notify
-
-selected_job = None
 
 
 def notify_job_info(state, var_name, value):
