+++ chapter 4/visual_elements/pages/scenario_element.py
@@ -1,5 +1,4 @@
 import taipy.gui.builder as tgb
-from orchestration import auto_scenario
 from taipy.gui import notify
